Исправлена ошибка, из-за которой в 64-х битных системах неправильно работал алгоритм кодирования/декодирования настроек.

+++ src/utils/options.cpp
@@ -118,8 +118,8 @@ QVariant stringToVariant(const QString &AString, QVariant::Type AType)
 #define rol(N, R) (((N) << (R)) | ((N) >> (32 - (R))))
 void xtea2_encipher(unsigned int num_rounds, quint32 *v, quint32 const *k)
 {
-	unsigned int i;
-	unsigned long a, b, c, d, sum=0, t,delta=0x9E3779B9;
+	quint32 i;
+	quint32 a, b, c, d, sum=0, t,delta=0x9E3779B9;
 	a = v[0];
 	b = v[1] + k[0];
 	c = v[2];
@@ -143,8 +143,8 @@ void xtea2_encipher(unsigned int num_rounds, quint32 *v, quint32 const *k)
 
 void xtea2_decipher(unsigned int num_rounds, quint32 *v, quint32 const *k)
 {
-	unsigned int i;
-	unsigned long a, b, c, d, t, delta=0x9E3779B9, sum=delta*num_rounds;
+	quint32 i;
+	quint32 a, b, c, d, t, delta=0x9E3779B9, sum=delta*num_rounds;
 	d = v[3];
 	c = v[2] ^ k[3];
 	b = v[1];
